requirements.txt
requests~=2.28.1
cohere~=2.9.1
annoy~=1.17.1
streamlit~=1.13.0
pandas~=1.3.5
pydub~=0.25.1
python-dotenv~=0.21.0